refactor(core): 重命名索引 verifyDataBase 文件为 manager & 更新导入
将 src/core/initialization/scoreBase/verifyDataBase.ts 重命名为 src/core/pid/manager.ts

同步更新 `main.ts` 的导入
