+++ src/core/pid/manager.ts
